+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project are documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [0.4.7] - Unreleased
+## [0.4.7] - 2025-07-08
 
 ### Added
 
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 
+- Comment Block formating `*/` now wraps to a new line.
 - Parser reporting `Unknown Syntax` in some exeptional cases for no apparent reason.
 - Parser support negative values in expression e.g.